fix: confirm page metadata
